refactor: rename EditingFeature to EditDialog
