+++ src/components/Atoms/Buttons/EditQuestionButton.js
@@ -1,7 +1,7 @@
 import React, {useState} from "react";
 import {IconButton, Tooltip} from "@material-ui/core";
 import EditIcon from "@material-ui/icons/Edit";
-import EditDialog2 from "../../Organisms/EditingFeature";
+import EditDialog2 from "../../Organisms/EditDialog";
 import store from "../../../store";
 import {SET_SAVED} from "../../../features/State Management/utilitiesSlice";
